Unwrap various redundant type casts

+++ upload/src/addons/SV/PasswordTools/Job/PasswordCleanup.php
@@ -9,7 +9,7 @@ class PasswordCleanup extends AbstractJob
 {
     public function run($maxRunTime): JobResult
     {
-        $pwnedPasswordCacheTime = (int)(\XF::options()->svPwnedPasswordCacheTime ?? 7);
+        $pwnedPasswordCacheTime = \XF::options()->svPwnedPasswordCacheTime ?? 7;
         $this->doCleanup($pwnedPasswordCacheTime);
 
         return $this->complete();
